+++ libmodernize/fixes/fix_filter.py
@@ -6,7 +6,7 @@ from __future__ import generator_stop
 from fissix import fixer_util
 from fissix.fixes import fix_filter
 
-import libmodernize
+from .. import utils
 
 
 class FixFilter(fix_filter.FixFilter):
@@ -15,7 +15,7 @@ class FixFilter(fix_filter.FixFilter):
 
     def transform(self, node, results):
         result = super().transform(node, results)
-        if not libmodernize.is_listcomp(result):
+        if not utils.is_listcomp(result):
             # Keep performance improvement from six.moves.filter in iterator
             # contexts on Python 2.7.
             fixer_util.touch_import("six.moves", "filter", node)
