+++ libmodernize/fixes/fix_map.py
@@ -6,7 +6,7 @@ from __future__ import generator_stop
 from fissix import fixer_util
 from fissix.fixes import fix_map
 
-import libmodernize
+from .. import utils
 
 
 class FixMap(fix_map.FixMap):
@@ -15,7 +15,7 @@ class FixMap(fix_map.FixMap):
 
     def transform(self, node, results):
         result = super().transform(node, results)
-        if not libmodernize.is_listcomp(result):
+        if not utils.is_listcomp(result):
             # Always use the import even if no change is required so as to have
             # improved performance in iterator contexts even on Python 2.7.
             fixer_util.touch_import("six.moves", "map", node)
