+++ libmodernize/fixes/fix_print.py
@@ -2,11 +2,11 @@ from __future__ import generator_stop
 
 from fissix.fixes import fix_print
 
-import libmodernize
+from .. import utils
 
 
 class FixPrint(fix_print.FixPrint):
     def transform(self, node, results):
         result = super().transform(node, results)
-        libmodernize.add_future(node, "print_function")
+        utils.add_future(node, "print_function")
         return result
