+++ libmodernize/fixes/fix_unicode_future.py
@@ -2,12 +2,12 @@ from __future__ import generator_stop
 
 from fissix.fixes import fix_unicode
 
-import libmodernize
+from .. import utils
 
 
 class FixUnicodeFuture(fix_unicode.FixUnicode):
     def transform(self, node, results):
         res = super().transform(node, results)
         if res:
-            libmodernize.add_future(node, "unicode_literals")
+            utils.add_future(node, "unicode_literals")
         return res
